git-svn-id: https://svn.origo.ethz.ch/zoodb/trunk@131 b278144e-e56a-4748-98cb-974f122dab01

+++ src/org/zoodb/jdo/internal/DataDeSerializer.java
@@ -595,7 +595,7 @@ public class DataDeSerializer {
     }
 
     private final Class<?> readClassInfo() {
-        final short id = _in.readShort();
+        final byte id = _in.readByte();
         if (id == -1) {
             //null-reference
             return null;
